requirements.txt
asgiref==3.3.1
Django==3.1.4
django-filter==2.4.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
future==0.18.2
Markdown==3.3.3
phonenumbers==8.12.13
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
serializers==0.2.4
sqlparse==0.4.1
